pull config into header.
