+++ livox_ros_driver/livox_ros_driver/lddc.h
@@ -47,19 +47,6 @@ namespace livox_ros
         kPclPxyziMsg
     } TransferType;
 
-    struct Lddc_config
-    {
-        int format;
-        int multi_topic;
-        int data_src;
-        int output_type;
-        double frequency;
-        std::string lidar_frame_id;
-        std::string imu_frame_id;
-        bool lidar_bag;
-        bool imu_bag;
-    };
-
     class Lddc
     {
     public:
